% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tauchart.r
\name{as_tauchart}
\alias{as_tauchart}
\title{Turn a simple (single-geom) ggplot plot into an tauchart object}
\usage{
as_tauchart(gg)
}
\arguments{
\item{gg}{ggplot object}
}
\value{
tauchart object
}
\description{
Takes a ggplot object that has a single geom (it can be geom_line,
geom_point or geom_histogram) and converts it to it's tauchart counterpart.
It will do it's best to identify plot labels, mapped size & color aesthetics,
and x/y limits.
}
\details{
If there are aesthetic mappings, \code{as_tauchart} will automaticlly add
a legend.

Currently supports:

\itemize{
  \item \code{geom_point}
  \item \code{geom_line}
  \item \code{geom_bar}
  \item \code{geom_histogram}
}
}
\note{
More aesthetic mappings are planned
}
\examples{
if (interactive()) {
library(ggplot2)
dat <- data.frame(year=seq(1790, 1970, 10),
                  uspop=as.numeric(uspop))
set.seed(5689)
data(movies, package="ggplot2")
movies <- movies[sample(nrow(movies), 1000), ]

gg <- ggplot(dat, aes(x=year, y=uspop)) + geom_line()
as_tauchart(gg)

gg <- ggplot(dat, aes(x=year, y=uspop)) + geom_point()
as_tauchart(gg)

gg <- ggplot(dplyr::count(movies, rating), aes(rating, n)) + geom_bar(stat="identity")
as_tauchart(gg)

gg <- ggplot(mtcars) + geom_point(aes(x=mpg, y=wt, color=cyl))
as_tauchart(gg)

gg <- ggplot(mtcars, aes(x=mpg, y=wt, color=am, size=wt)) + geom_point()
as_tauchart(gg)

data(economics, package="ggplot2")
gg <- ggplot(economics) + geom_line(aes(x=date, y=unemploy))
as_tauchart(gg) \%>\% tau_guide_x(tick_format="\%Y")

gg <- ggplot(mtcars, aes(as.factor(cyl))) + geom_histogram()
as_tauchart(gg)
}
}
