% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{run_tau_app}
\alias{run_tau_app}
\title{Run a built-in example Shiny app}
\usage{
run_tau_app(app_name = NULL)
}
\arguments{
\item{app_name}{name of built-in Shiny app to run (if \code{NULL} displays a list
of built-in apps)}
}
\description{
When run without specifying \code{app_name}, this function will return
a list of built-in TauCharts Shiny example applications. To run one of the
built-in example applications, specify only the name in the \code{app_name}
parameter and the function will display the full path to the app (making it
easier to find and view the source) and then execute the app.
}
\examples{
\dontrun{
# list the built-in apps

run_tau_app()
## Available shiny example apps:
## -----------------------------
## tau01 - TauCharts Shiny Example 01

# run an app

# run_tau_app("tau01")
## Running shiny app [tau01] from:
}
}
