% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.r
\name{tau_area}
\alias{tau_area}
\title{Create a TauCharts area chart}
\usage{
tau_area(tau, x, y, color = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{color}{quoted name of \code{data} column to map color aesthetic to}
}
\description{
Create a TauCharts area chart
}
\examples{
data(economics, package="ggplot2")
tauchart(economics) \%>\%
  tau_area("date", "unemploy") \%>\%
  tau_guide_x(tick_format="\%Y")

# facets
library(dplyr)
library(tidyr)

crimes <- gather(add_rownames(USArrests, "State"), Crime, Amount, -State)
tauchart(crimes) \%>\%
  tau_area("State", c("Crime", "Amount"), "Crime") \%>\%
  tau_guide_y(auto_scale = FALSE)
}
\references{
\url{http://api.taucharts.com/basic/bar.html},
            \url{http://api.taucharts.com/basic/horizontal-bar.html}
}

