% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taucharts.R
\name{tau_bar}
\alias{tau_bar}
\title{Create a TauCharts bar chart (horizontal or vertical)}
\usage{
tau_bar(tau, x, y, color = NULL, size = NULL, horizontal = FALSE)
}
\arguments{
\item{tau}{taucharts object}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{color}{quoted name of \code{data} column to map color aesthetic to}

\item{size}{quoted name of \code{data} column to make size aesthetic to}

\item{horizontal}{should the bar chart be horizontal? (default: \code{FALSE} (no))}
}
\description{
Create a TauCharts bar chart (horizontal or vertical)
}
\references{
\url{http://api.taucharts.com/basic/line.html}
}

