% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.r
\name{tau_boxplot}
\alias{tau_boxplot}
\title{Create a TauCharts box plot (horizontal or vertical)}
\usage{
tau_boxplot(tau, x, y, horizontal = FALSE, mode = "outliers-only")
}
\arguments{
\item{tau}{taucharts object}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{horizontal}{should the bar chart be horizontal? (default: \code{FALSE} (no))}

\item{mode}{when to show scatter points? e.g. "hide-scatter", "show-scatter", "outliers-only"}
}
\description{
There is no option for color mappings as the tauCharts plugin does not support this yet.
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_boxplot("gear", "mpg", mode="hide-scatter") \%>\%
  tau_guide_y(nice=F)
}
}
