% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_538}
\alias{tau_color_538}
\title{Use the "fivethirtyeight.com" palette used in the charts}
\usage{
tau_color_538(tau, n = 5)
}
\arguments{
\item{tau}{taucharts object}

\item{n}{number of desired colors}
}
\description{
The standard fivethirtyeight.com palette for line plots is blue, red, green.
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_538()
}
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
\seealso{
\code{\link[ggthemes]{fivethirtyeight_pal}}
}

