% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_brewer}
\alias{tau_color_brewer}
\title{Use the ColorBrewer palette in the charts}
\usage{
tau_color_brewer(tau, n = 5, palette = "Set2")
}
\arguments{
\item{tau}{taucharts object}

\item{n}{number of colors to generate (no checking is done to ensure the palette
has \code{n} colors).}

\item{palette}{ColorBrewer palette name. One of \code{"Set2"}, \code{"BrBG"},
\code{"PiYG"}, \code{"PRGn"}, \code{"PuOr"}, \code{"RdBu"}, \code{"RdGy"}, \code{"RdYlBu"},
\code{"RdYlGn"}, \code{"Spectral"}, \code{"Accent"}, \code{"Dark2"}, \code{"Paired"},
\code{"Pastel1"}, \code{"Pastel2"}, \code{"Set1"}, \code{"Set3"}, \code{"Blues"},
\code{"BuGn"}, \code{"BuPu"}, \code{"GnBu"}, \code{"Greens"}, \code{"Greys"}, \code{"Oranges"},
\code{"OrRd"}, \code{"PuBu"}, \code{"PuBuGn"}, \code{"PuRd"}, \code{"Purples"},
\code{"RdPu"}, \code{"Reds"}, \code{"YlGn"}, \code{"YlGnBu"}, \code{"YlOrBr"}, \code{"YlOrRd"}.}
}
\description{
Use the ColorBrewer palette in the charts
}
\note{
It is highly suggested that callers use the ColorBrewer qualitative palettes:
      ["Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2", "Set3"],
      especially if you are plotting categorical values (which you most liklely are
      since you're using this package).
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_brewer(3, "Set3")
}
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
