% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_economist}
\alias{tau_color_economist}
\title{Use the "Economist" palette used in the charts}
\usage{
tau_color_economist(tau, n = 5)
}
\arguments{
\item{tau}{taucharts object}

\item{n}{number of desired colors}
}
\description{
The hues in the palette are blues, grays, and greens. Red is not included in
these palettes and should be used to indicate important data.
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_economist()
}
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
\seealso{
\code{\link[ggthemes]{economist_pal}}
}

