% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_few}
\alias{tau_color_few}
\title{Use the "Few" palette used in the charts}
\usage{
tau_color_few(tau, n = NULL, palette = "medium")
}
\arguments{
\item{tau}{taucharts object}

\item{n}{number of desired colors}

\item{palette}{name of Few palette. One of \code{"medium"}, \code{"dark"} or \code{"light"}}
}
\description{
Qualitative color palettes from Stephen Few, "Practical Rules for Using Color in Charts".
}
\examples{
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_few()
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
\seealso{
\code{\link[ggthemes]{few_pal}}
}
