% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_highcharts}
\alias{tau_color_highcharts}
\title{Use the HighchartsJS palette used in the charts}
\usage{
tau_color_highcharts(tau, n = 5, palette = "default")
}
\arguments{
\item{tau}{taucharts object}

\item{n}{number of desired colors}

\item{palette}{name of Few palette. One of \code{"default"} or \code{"darkunica"}}
}
\description{
Qualitative color palettes from Stephen Few, "Practical Rules for Using Color in Charts".
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_highcharts()
}
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
\seealso{
\code{\link[ggthemes]{hc_pal}}
}

