% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_manual}
\alias{tau_color_manual}
\title{Specify the colors used in the charts}
\usage{
tau_color_manual(tau, values = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{values}{vector of colors, ideally names (e.g. "\code{black}") or
hex-format (e.g. "\code{#ffeeaa}")}
}
\description{
Vector can be optionally named in order to give explicit color/value mapping.
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_manual(c("blue", "maroon", "black"))
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_manual(c(`4`="blue",`6`= "maroon",`8`= "black"))
}
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
