% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_tableau}
\alias{tau_color_tableau}
\title{Use the Tableau palette in the charts}
\usage{
tau_color_tableau(tau, palette = "tableau20")
}
\arguments{
\item{tau}{taucharts object}

\item{palette}{Tableau palette name. One of \code{"tableau20"},
\code{"tableau10medium"}, \code{"gray5"}, \code{"colorblind10"},
\code{"purplegray12"}, \code{"bluered12"}, \code{"greenorange12"}, \code{"cyclic"}.}
}
\description{
Use the Tableau palette in the charts
}
\examples{
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_tableau()
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
