% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_wsj}
\alias{tau_color_wsj}
\title{Use the "Wall Street Journal" palette used in the charts}
\usage{
tau_color_wsj(tau, n = 5, palette = "rgby")
}
\arguments{
\item{tau}{taucharts object}

\item{n}{number of desired colors}

\item{palette}{name of Few palette. One of \code{"rgby"}, \code{"green_red"},
\code{"green_black"}, \code{"dem_rep"} or \code{"colors6"}}
}
\description{
A subset of the plethora of palettes used by the WSJ.
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("wt", "mpg", color="cyl") \%>\%
  tau_color_wsj()
}
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}
\seealso{
\code{\link[ggthemes]{wsj_pal}}
}

