% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_export_plugin}
\alias{tau_export_plugin}
\title{Add a TauCharts export plugin}
\usage{
tau_export_plugin(tau)
}
\arguments{
\item{tau}{taucharts object}
}
\description{
Add a TauCharts export plugin
}
\examples{
data(cars_data)
tauchart(cars_data) \%>\%
  tau_point("milespergallon", c("class", "price"), color="class") \%>\%
  tau_quick_filter()
}
\seealso{
\code{\link{cars_data}} dataset
}

