% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guides.r
\name{tau_guide_padding}
\alias{tau_guide_padding}
\title{Set overall chart padding}
\usage{
tau_guide_padding(tau, bottom = 0, left = 0, top = 0, right = 0)
}
\arguments{
\item{tau}{taucharts object}

\item{bottom,top,left,right}{amount of padding}
}
\description{
Set overall chart padding
}
\examples{
tauchart(mtcars) \%>\%
tau_point("mpg", "wt") \%>\%
  tau_guide_padding(40, 40, 40, 40)
}
\references{
\url{http://api.taucharts.com/basic/guide.html}
}

