% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.r
\name{tau_guide_x}
\alias{tau_guide_x}
\title{Control x-axis padding, label, scale & tick format}
\usage{
tau_guide_x(tau, padding = NULL, label = NULL, label_padding = NULL,
  nice = NULL, auto_scale = NULL, tick_period = NULL,
  tick_format = NULL, min = NULL, max = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{padding}{space between axis ticks and chart panel}

\item{label}{text of label for axis (overrides default use of variable name)}

\item{label_padding}{space between axis ticks and axis label (can be negative)}

\item{nice}{Taucharts engine tries to make axis scale "nice" by trying to start measure-based scale from 0 and adds some margins to complete scale with "nice" numbers. For example, if original scale domain contains values [8, 20, ... 40], then axis will have ticks from 0 to 45.
(default: \code{TRUE}, if min and max values are set: \code{FALSE}).}

\item{auto_scale}{(Deprecated) auto-pick "best" scale for axis? (default: \code{TRUE} (yes))}

\item{tick_period}{if axis is auto-determined to be a "period scale",
this allows specification of the period size. See \code{References} for more
information.}

\item{tick_format}{can be any
\href{https://github.com/mbostock/d3/wiki/Formatting#d3_format}{D3 format specifier}}

\item{min, max}{manual minimum and maximum for the axis}
}
\description{
Control x-axis padding, label, scale & tick format
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
 tau_point("mpg", "wt") \%>\%
 tau_guide_x(label="Miles/gallon", nice=FALSE) \%>\%
 tau_guide_y(label="Weight", nice=FALSE)
}
}
\references{
\url{http://api.taucharts.com/basic/guide.html},
            \url{https://github.com/mbostock/d3/wiki/Formatting#d3_format}
}
