% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_legend}
\alias{tau_legend}
\title{Add a TauCharts legend}
\usage{
tau_legend(tau)
}
\arguments{
\item{tau}{taucharts object}
}
\description{
Add a TauCharts legend
}
\examples{
data(cars_data)
tauchart(cars_data) \%>\%
  tau_point("milespergallon", c("class", "price"), color="class") \%>\%
  tau_legend()
}
\seealso{
\code{\link{cars_data}} dataset
}
