% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taucharts.R
\name{tau_line}
\alias{tau_line}
\title{Create a TauCharts line chart}
\usage{
tau_line(tau, data, x, y, color = NULL, size = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{data}{data.frame}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{color}{quoted name of \code{data} column to map color aesthetic to}

\item{size}{quoted name of \code{data} column to make size aesthetic to}
}
\description{
Create a TauCharts line chart
}
\references{
\url{http://api.taucharts.com/basic/bar.html},
            \url{http://api.taucharts.com/basic/horizontal-bar.html}
}

