% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_quick_filter}
\alias{tau_quick_filter}
\title{Add a TauCharts quick filter}
\usage{
tau_quick_filter(tau)
}
\arguments{
\item{tau}{taucharts object}

\item{hideError}{\code{logical} to show errors.}
}
\description{
Add a TauCharts quick filter
}
\examples{
data(cars_data)
tauchart(cars_data) \%>\%
  tau_point("milespergallon", c("class", "price"), color="class") \%>\%
  tau_quick_filter()
}
\seealso{
\code{\link{cars_data}} dataset
}

