% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonts.r
\name{tau_set_font}
\alias{tau_set_font}
\title{Set the font family for the TauCharts chart}
\usage{
tau_set_font(tau,
  family = "OpenSans, 'Helvetica Neue', Helvetica, Arial, sans-serif",
  import_url = "https://fonts.googleapis.com/css?family=Open+Sans:400,400italic,600,700,600italic,700italic,800,300,300italic,800italic")
}
\arguments{
\item{family}{CSS \code{font-family} spec (just the comma-separated font names)}

\item{import_url}{the \code{<link>} href if the fonts come from an external source}
}
\value{
taucharts object
}
\description{
Set the font family (standard CSS \code{font-family} specification, comma-separated
font names only) and an optional font CSS URL to load the font from.
}
\examples{
tauchart(mtcars) \%>\%
  tau_point("mpg", "wt", "cyl") \%>\%
  tau_legend() \%>\%
  tau_color_brewer() \%>\%
  tau_tooltip() \%>\%
  tau_set_font("'Montserrat', sans-serif", "http://fonts.googleapis.com/css?family=Montserrat:400,700")
}

