% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.r
\name{tau_stacked_bar}
\alias{tau_stacked_bar}
\title{Create a TauCharts stacked bar chart (experimental)}
\usage{
tau_stacked_bar(tau, x, y, color = NULL, size = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{color}{quoted name of \code{data} column to map color aesthetic to.
NOTE that the parameter to this is what really defines the stacking.}

\item{size}{quoted name of \code{data} column to make size aesthetic to}
}
\description{
The API supports it but it's not documented at all
}
\examples{
data(mpg, package="ggplot2")
tauchart(dplyr::count(mpg, class, drv)) \%>\%
  tau_stacked_bar("class", "n", "drv") \%>\%
  tau_guide_gridlines(FALSE, FALSE) \%>\%
  tau_tooltip()
}
\references{
\url{http://api.taucharts.com/basic/line.html}
}

