% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tasks.r
\name{tau_tasks}
\alias{tau_tasks}
\title{Add post-render JavaScript tasks to taucharts}
\usage{
tau_tasks(tau, task = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{task}{either a \code{character} or \code{\link[htmlwidgets]{JS}}
         representing a JavaScript function to run after a tauchart has
         been rendered.  This provides the ability for advanced customization.
         The JavaScript function will be \code{call}ed so \code{this.el} will be
         the containing \code{div} element and \code{this.chart} will be the
         \code{taucharts} object.}
}
\description{
Add post-render JavaScript tasks to taucharts
}
\examples{
library(taucharts)

tauchart(mtcars) \%>\%
  tau_point("mpg", "wt") \%>\%
  tau_tasks("function(){alert('I drew a chart')}") \%>\%
  tau_tasks(htmlwidgets::JS(
    "function(){alert('Yep, I really did.')}"
  )) \%>\%
  tau_tasks(
    "
function(){
  d3.select(this.el).selectAll('.graphical-report__dot')
    .transition().style('opacity',0.1)
    .transition().style('opacity',0.9)
    .transition().style('opacity',0.1)
    .transition().style('opacity',0.9);
}
"
  )
}

