% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_tooltip}
\alias{tau_tooltip}
\title{Add a TauCharts tooltip}
\usage{
tau_tooltip(tau, fields = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{fields}{\code{vector} of fields you would like to see in the tooltip.
If \code{NULL} then will assume all fields.}
}
\description{
Add a TauCharts tooltip
}

