% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_trendline}
\alias{tau_trendline}
\title{Add a TauCharts trendline}
\usage{
tau_trendline(tau, type = NULL, hideError = FALSE, showPanel = TRUE,
  showTrend = TRUE, models = c("linear", "exponential", "logarithmic",
  "loess", "lastvalue", "polynomial", "power"))
}
\arguments{
\item{tau}{taucharts object}

\item{type}{\code{character} Model representing default trend line to show. Must be
one of models specified in models parameter.
If unspecified, will use the first model specified in models.}

\item{hideError}{\code{logical} to show errors.}

\item{showPanel}{\code{logical} to show the panel next to the chart to allow a user
to manipulate the trendlines.  When \code{FALSE}, the trendlines will
still appear though.}

\item{showTrend}{\code{logical} to show the trendlines on initial display.  If
\code{showPanel = TRUE}, then the user will have the opportunity
to add/delete the trendlines.}

\item{models}{\code{character} or \code{vector} of \code{characters} for the models
to show in the trendline panel if \code{showPanel = TRUE}. If you
would like to change the order of the options, then you can do
\code{models = c("logarithmic","exponential")}, and the first provided
will be the initial model type used.}
}
\description{
Add a TauCharts trendline
}
\examples{
data(cars_data)
tauchart(cars_data) \%>\%
  tau_point("milespergallon", c("class", "price"), color="class") \%>\%
  tau_trendline()
}
\seealso{
\code{\link{cars_data}} dataset
}

