% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taucharts-package.r
\name{taucharts-exports}
\alias{JS}
\alias{\%>\%}
\alias{saveWidget}
\alias{taucharts-exports}
\title{taucharts exported operators}
\description{
The following functions are imported and then re-exported
from the taucharts package to enable use of the magrittr
pipe operator with no additional library calls
}

