% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taucharts-package.r
\docType{package}
\name{taucharts-package}
\alias{taucharts-package}
\title{An \code{htmlwidget} interface to the
\href{http://www.taucharts.com/}{TauCharts} D3 chart library}
\description{
An \code{htmlwidget} interface to the
\href{http://www.taucharts.com/}{TauCharts} D3 chart library
}
\author{
Bob Rudis (@hrbrmstr)
}

