% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taucharts-package.r, R/taucharts.R
\docType{package}
\name{taucharts}
\alias{taucharts}
\alias{taucharts-package}
\title{An \code{htmlwidget} interface to the
\href{http://www.taucharts.com/}{TauCharts} D3 chart library}
\usage{
taucharts(width = NULL, height = NULL)
}
\arguments{
\item{width,height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\description{
An \code{htmlwidget} interface to the
\href{http://www.taucharts.com/}{TauCharts} D3 chart library

Performs basic widget setup and returns an object suitable for
use with the other \code{tau_} functions.
}
\author{
Bob Rudis (@hrbrmstr)
}
\references{
\url{http://api.taucharts.com/}
}

