% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.r
\name{tau_bar}
\alias{tau_bar}
\title{Create a TauCharts bar chart (horizontal or vertical)}
\usage{
tau_bar(tau, x, y, color = NULL, size = NULL, horizontal = FALSE)
}
\arguments{
\item{tau}{taucharts object}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{color}{quoted name of \code{data} column to map color aesthetic to}

\item{size}{quoted name of \code{data} column to make size aesthetic to}

\item{horizontal}{should the bar chart be horizontal? (default: \code{FALSE} (no))}
}
\description{
Create a TauCharts bar chart (horizontal or vertical)
}
\examples{
data(mpg, package="ggplot2")
tauchart(dplyr::count(mpg, class)) \%>\%
  tau_bar("class", "n")

# facets
tauchart(dplyr::count(mpg, manufacturer, class)) \%>\%
  tau_bar("class", c("manufacturer", "n"))
#facets
mfclass <- dplyr::count(mpg, manufacturer, model, class)
tauchart(mfclass) \%>\%
  tau_bar(c("manufacturer", "model"), "n")

# ordered factors on x-axis
mfclass$class <- factor(mfclass$class,
                        levels=c("2seater", "subcompact", "compact",
                                 "midsize", "minivan",  "suv", "pickup"),
                        ordered=TRUE)
tauchart(mfclass) \%>\%
  tau_bar(c("class", "manufacturer"), "n")
}
\references{
\url{http://api.taucharts.com/basic/line.html}
}

