% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colors.r
\name{tau_color_manual}
\alias{tau_color_manual}
\title{Specify the colors used in the charts}
\usage{
tau_color_manual(tau, values = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{vector}{of colors, ideally names (e.g. "\code{black}") or
hex-format (e.g. "\code{#ffeeaa}")}
}
\description{
Specify the colors used in the charts
}
\references{
\url{http://api.taucharts.com/advanced/encoding.html}
}

