% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.r
\name{tau_guide_gridlines}
\alias{tau_guide_gridlines}
\title{Control showing of axis gridlines}
\usage{
tau_guide_gridlines(tau, show_x = TRUE, show_y = TRUE)
}
\arguments{
\item{tau}{taucharts object}

\item{show_x, show_y}{if \code{TRUE}, show the gridline}
}
\description{
Control showing of axis gridlines
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\%
  tau_point("mpg", "wt") \%>\%
  tau_guide_gridlines(FALSE, FALSE)
}
}
\references{
\url{http://api.taucharts.com/basic/guide.html}
}

