% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.r
\name{tau_point}
\alias{tau_point}
\title{Create a TauCharts scatterplot}
\usage{
tau_point(tau, x, y, color = NULL, size = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{x}{quoted name of \code{data} column to use for x-axis values}

\item{y}{quoted name of \code{data} column to use for y-axis values}

\item{color}{quoted name of \code{data} column to map color aesthetic to}

\item{size}{quoted name of \code{data} column to make size aesthetic to}
}
\description{
Create a TauCharts scatterplot
}
\examples{
tauchart(mtcars) \%>\% tau_point("mpg", "wt")
mtcars$cyl <- factor(mtcars$cyl)
tauchart(mtcars) \%>\% tau_point("mpg", "wt", "cyl", "hp")
}
\references{
\url{http://api.taucharts.com/basic/scatterplot.html}
}

