% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_quick_filter}
\alias{tau_quick_filter}
\title{Add a TauCharts quick filter plugin}
\usage{
tau_quick_filter(tau, fields = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{fields}{fields which should be shown in quick filter}
}
\description{
Add a TauCharts quick filter plugin
}
\examples{
data(cars_data)
tauchart(cars_data) \%>\%
  tau_point("milespergallon", c("class", "price"), color="class") \%>\%
  tau_quick_filter(fields = c("price"))
}
\seealso{
\code{\link{cars_data}} dataset
}

