% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugins.R
\name{tau_tooltip}
\alias{tau_tooltip}
\title{Add a TauCharts tooltip}
\usage{
tau_tooltip(tau, fields = NULL)
}
\arguments{
\item{tau}{taucharts object}

\item{fields}{character vector of fields to display in the tooltip
(default is to use all columns in \code{data})}
}
\description{
Add a TauCharts tooltip
}
\examples{
if (interactive()) {
data(cars_data)
tauchart(cars_data) \%>\%
  tau_point("milespergallon", c("class", "price"), color="class") \%>\%
  tau_tooltip(c("vehicle", "year", "class", "price", "milespergallon"))
}
}
\seealso{
\code{\link{cars_data}} dataset
}

