% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taucharts.R
\name{tauchart}
\alias{tauchart}
\title{Create a new TauChart}
\usage{
tauchart(data, width = NULL, height = NULL, inputId = NULL)
}
\arguments{
\item{data}{\code{data.frame}-like or \code{xts} object}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{inputId}{the input slot used to access the clicked point (for Shiny).}
}
\description{
Performs basic widget setup and returns an object suitable for
use with the other \code{tau_} functions.
}
\examples{
if (interactive()) {
tauchart(mtcars) \%>\% tau_point("mpg", "wt")
}
}
\references{
\url{http://api.taucharts.com/}
}
