% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taucharts.R
\name{taucharts}
\alias{taucharts}
\title{Create a new TauChart}
\usage{
taucharts(width = NULL, height = NULL)
}
\arguments{
\item{width,height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\description{
Performs basic widget setup and returns an object suitable for
use with the other \code{tau_} functions.
}
\references{
\url{http://api.taucharts.com/}
}

